Move error constants to hardwarekey package.
